Add test using mockito and jacoco report and update ids to be integer

+++ src/main/java/com/retaildiscountssvc/demo/api/dao/ItemRepository.java
@@ -3,5 +3,5 @@ package com.retaildiscountssvc.demo.api.dao;
 import com.retaildiscountssvc.demo.api.model.Item;
 import org.springframework.data.mongodb.repository.MongoRepository;
 
-public interface ItemRepository extends MongoRepository<Item, Long> {
+public interface ItemRepository extends MongoRepository<Item, Integer> {
 }
